Add new API for PageView implementation in SwiftUI
The new API replaces the PagingItem initializer with a new Page struct
that let's you define custom header items per page. Each page can
define their own custom SwiftUI view as the menu item, and it can be
customized based on the selection and progress. The new API also
supports result builders, which makes it easier to create PageViews
with a fixed number of pages.