Fix PageViewController isScrollingFrom delegate when overshooting
When swiping fast to the next view controller, the scroll view will
overshoot and trigger additional isScrollingFrom methods after
didFinishScrollingFrom is called. This is expected, but the view
controllers that are returned are the previously selected view
controller and not the upcoming view controller. This was done to
avoid triggered view controller appearance method (like
viewWillAppear) for the upcoming view controllers when overshooting,
but it seems like this is not happening anymore, so we can just remove
this workaround.