Fix issues with SwiftUI views being reset when updating
The current implementation calls reloadData on each update, which
re-creates the entire content view and resets any state (like
scrolling). To prevent this, we update the existing
UIHostingController if the current page is the same.

This requires us to have stable identifiers, to know if the current
page is the same as before or a new one. For title-based pages, we
just use the title as the identifier. For pages with custom SwiftUI
headers, we default to using the index of the view as the
identifier. This is the same behaviour that we have today, although it
will only work when having static pages. In order to support dynamic
pages with custom SwiftUI headers, we introduce a new initializer that
allows specifying the identifier.